test: 애플 소셜 로그인 기능 UseCase 테스트 코드 추가
#27
